requirements.txt
annotated-types==0.6.0
cheshire_cat_api==1.4.1
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
websocket-client==1.7.0
numpy